creating methods in StatisticService file and Update StatisticRequestDto file

+++ desafio-itaus/src/main/java/com/mak/desafio_itaus/controller/dtos/StatisticRequestDto.java
@@ -1,6 +1,6 @@
 package com.mak.desafio_itaus.controller.dtos;
 
-public record StatisticRequestDto(Integer count,
+public record StatisticRequestDto(Long count,
                                   Double sum,
                                   Double avg,
                                   Double min,
